#77 - start restructoring app - really big commit
