+++ server/src/app/index.ts
@@ -1,0 +1,3 @@
+export * from "./interfaces/IBase.repository";
+export * from "./models/BaseObject";
+export * from "./repositories/Base.repository";
