+++ server/src/app/interfaces/IBase.repository.ts
