+++ server/src/app/models/BaseObject.ts
