+++ server/src/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./handler/auth-validation.handler";
+export * from "./handler/auth.handler";
