+++ server/src/db/repositories/index.ts
@@ -1,7 +1,1 @@
 export * from "./recipe.repository";
-export * from "./IBase.repository";
-export * from "./Base.repository";
-export * from "./SocoboUser.repository";
-export * from "./SocoboUserRole.repository";
-export * from "./SocoboUserProvider.repository";
-export * from "./SocoboUserImage.repository";
