+++ server/src/logic/handler/index.ts
@@ -1,9 +1,3 @@
-export * from "./validation/auth-validation.handler";
 export * from "./validation/model-validation.handler";
-export * from "./routes/auth.handler";
 export * from "./routes/log.handler";
 export * from "./routes/recipe.handler";
-export * from "./routes/socobouser.handler";
-export * from "./routes/socobouserImages.handler";
-export * from "./routes/socobouserProviders.handler";
-export * from "./routes/socobouserRoles.handler";
