--- server/src/models/enums/provider-type.ts
@@ -1,4 +1,0 @@
-export enum ProviderType {
-  Email = 1,
-  Username = 2
-}
