--- server/src/models/enums/role.ts
@@ -1,4 +1,0 @@
-export enum Role {
-  Admin = 1,
-  User = 2
-}
