+++ server/src/models/index.ts
@@ -1,14 +1,7 @@
 export * from "./database/recipe";
 export * from "./database/recipe-step";
 export * from "./database/food-item";
-export * from "./database/SocoboUser";
-export * from "./database/SocoboUserImage";
-export * from "./database/SocoboUserProvider";
-export * from "./database/SocoboUserRole";
 export * from "./enums/data-types";
-export * from "./enums/provider-type";
-export * from "./enums/role";
-export * from "./enums/update-type";
 export * from "./enums/validation-group";
 export * from "./errors/api-error";
 export * from "./errors/db-error";
