+++ server/src/routes/api/v1/index.ts
@@ -1,7 +1,3 @@
 export * from "./auth/auth.route";
 export * from "./logs/log.route";
 export * from "./recipes/recipe.route";
-export * from "./socobousers/socobouserImages.route";
-export * from "./socobousers/socobouserProviders.route";
-export * from "./socobousers/socobouserRoles.route";
-export * from "./socobousers/socobousers.route";
