+++ server/src/socobouser/enums/SocoboUserProviderTypes.ts
@@ -1,0 +1,4 @@
+export enum SocoboUserProviderTypes {
+  Email = 1,
+  Username = 2
+}
