+++ server/src/socobouser/enums/SocoboUserRoleTypes.ts
@@ -1,0 +1,4 @@
+export enum SocoboUserRoleTypes {
+  Admin = 1,
+  User = 2
+}
