+++ server/src/socobouser/handlers/SocoboUserImages.handler.ts
@@ -1,8 +1,10 @@
 import { Request, Response } from "express";
 import { IDatabase } from "pg-promise/typescript/pg-promise";
-import { DbExtensions, SocoboUserRole } from "../../../models/index";
+import { DbExtensions } from "../../models/index";
+import { SocoboUserImage } from "../models/SocoboUserImage";
 
-export class SocoboUserRolesHandler {
+export class SocoboUserImagesHandler {
+
   private _db: IDatabase<DbExtensions>&DbExtensions;
 
   constructor (db: any) {
@@ -10,29 +12,29 @@ export class SocoboUserRolesHandler {
   }
 
   public getAll = (req: Request, res: Response): void => {
-    this._db.socobouserRoles.getAll()
-      .then((result: SocoboUserRole[]) => res.status(200).json(result))
+    this._db.socobouserImages.getAll()
+      .then((result: SocoboUserImage[]) => res.status(200).json(result))
       .catch((e: any) => res.status(e.statusCode).json(e.forResponse()));
   }
 
   public getById = (req: Request, res: Response): void => {
-    const roleId: number = +req.params.id;
-    this._db.socobouserRoles.getById(roleId)
-      .then((result: SocoboUserRole) => res.status(200).json(result))
+    const imageId: number = +req.params.id;
+    this._db.socobouserImages.getById(imageId)
+      .then((result: SocoboUserImage) => res.status(200).json(result))
       .catch((e: any) => res.status(e.statusCode).json(e.forResponse()));
   }
 
   public updateById = (req: Request, res: Response): void => {
-    const roleId: number = +req.params.id;
+    const imageId: number = +req.params.id;
     const fieldValue: string = req.body.fieldValue;
-    this._db.socobouserRoles.updateById(roleId, fieldValue)
-      .then((result: SocoboUserRole) => res.status(200).json(result))
+    this._db.socobouserImages.updateById(imageId, fieldValue)
+      .then((result: SocoboUserImage) => res.status(200).json(result))
       .catch((e: any) => res.status(e.statusCode).json(e.forResponse()));
   }
 
   public deleteById = (req: Request, res: Response): void => {
-    const roleId: number = +req.params.id;
-    this._db.socobousers.deleteById(roleId)
+    const imageId: number = +req.params.id;
+    this._db.socobouserImages.deleteById(imageId)
       .then((result: Object) => res.status(200).json(result))
       .catch((e: any) => res.status(e.statusCode).json(e.forResponse()));
   }
