+++ server/src/socobouser/models/SocoboUserRole.ts
@@ -1,6 +1,6 @@
 import { IsNotEmpty, MinLength } from "class-validator";
-import { ValidationGroup } from "./../enums/validation-group";
-import { BaseObject } from "./BaseObject";
+import { BaseObject } from "../../base/index";
+import { ValidationGroup } from "../../models/index";
 
 export class SocoboUserRole extends BaseObject {
