+++ server/src/socobouser/repositories/SocoboUserProvider.repository.ts
@@ -1,15 +1,15 @@
-import { SocoboUserRole } from "../../models/index";
-import { BaseRepository } from "./Base.repository";
+import { SocoboUserProvider } from "../models/SocoboUserProvider";
+import { BaseRepository } from "../../base/index";
 
-export class SocoboUserRoleRepository extends BaseRepository <SocoboUserRole> {
+export class SocoboUserProviderRepository extends BaseRepository <SocoboUserProvider> {
 
   constructor (db: any) {
-    super(db, "Socobo_User_Role", "name");
+    super(db, "Socobo_User_Provider", "name");
     this.transformFunction = this._transformResult;
   }
 
-  private _transformResult = (result: any): SocoboUserRole => {
-    const tranformedResult: SocoboUserRole = new SocoboUserRole()
+  private _transformResult = (result: any): SocoboUserProvider => {
+    const tranformedResult: SocoboUserProvider = new SocoboUserProvider()
       .addId(result.hasOwnProperty("id") ? Number(result.id) : null)
       .addName(result.hasOwnProperty("name") ? result.name : null)
       .addCreated(result.hasOwnProperty("created") ? Number(result.created) : null)
