+++ server/src/socobouser/routes/SocoboUserImages.route.ts
@@ -1,30 +1,31 @@
 import { Router } from "express";
-import { AuthValidationHandler, SocoboUserProvidersHandler } from "./../../../../logic/handler/index";
+import { SocoboUserImagesHandler } from "../handlers/SocoboUserImages.handler";
+import { AuthValidationHandler } from "../../logic/handler/index";
 
-export class SocoboUserProvidersRoute {
+export class SocoboUserImagesRoute {
 
   constructor (
     private _router: Router,
-    private _socoboUserProvidersHandler: SocoboUserProvidersHandler,
+    private _socoboUserImagesHandler: SocoboUserImagesHandler,
     private _authValidationHandler: AuthValidationHandler
   ) {}
 
   public createRoutes (): Router {
     this._router.get("/",
       this._authValidationHandler.checkToken,
-      this._socoboUserProvidersHandler.getAll);
+      this._socoboUserImagesHandler.getAll);
 
     this._router.get("/:id",
       this._authValidationHandler.checkToken,
-      this._socoboUserProvidersHandler.getById);
+      this._socoboUserImagesHandler.getById);
 
     this._router.put("/:id",
       this._authValidationHandler.checkToken,
-      this._socoboUserProvidersHandler.updateById);
+      this._socoboUserImagesHandler.updateById);
 
     this._router.delete("/:id",
       this._authValidationHandler.checkToken,
-      this._socoboUserProvidersHandler.deleteById);
+      this._socoboUserImagesHandler.deleteById);
 
     return this._router;
   }
